refactoring query summary message out of QuerySummary
https://coveord.atlassian.net/browse/JSUI-2237
